+++ crates/client/rpc/src/methods/read/call.rs
@@ -3,8 +3,8 @@ use crate::errors::StarknetRpcResult;
 use crate::methods::trace::trace_transaction::FALLBACK_TO_SEQUENCER_WHEN_VERSION_BELOW;
 use crate::Starknet;
 use dc_exec::ExecutionContext;
-use starknet_core::types::Felt;
 use starknet_core::types::{BlockId, FunctionCall};
+use starknet_types_core::felt::Felt;
 use std::sync::Arc;
 
 /// Call a Function in a Contract Without Creating a Transaction